pollycracker/requirements.txt
PyCryptodome
colorama